pyproject.toml
[tool.poetry]
name = "chategram"
version = "0.1.0"
description = ""
authors = ["Anders Aagaard <aagaande@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^0.27.7"
python-telegram-bot = "^20.3"
guidance = "^0.0.57"
python-dotenv = "^1.0.0"
trafilatura = "^1.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
